+++ app/utils/validators/CheckBoardOwnerValidator.java
@@ -4,10 +4,10 @@ import models.User;
 import play.data.validation.Constraints;
 import play.libs.F;
 
-public class CheckBoardOwnerValidator extends Constraints.Validator<Integer> {
+public class CheckBoardOwnerValidator extends Constraints.Validator<Long> {
 
     @Override
-    public boolean isValid(Integer boardID) {
+    public boolean isValid(Long boardID) {
         return boardID == null ||
                 (User.isLoggedIn() &&
                         User.loggedInUser().boards.stream().anyMatch(board -> board.id.equals(boardID)));
